renamed upsert-election-doc to find-and-set-election-doc
